Add unit tests for FlightService, FlightScheduleService, FlightInstanceService, AircraftManagementService

+++ ams-airline-manag-system/src/main/java/com/ra/course/ams/airline/manag/system/service/impl/FlightInstanceServiceImpl.java
@@ -7,6 +7,7 @@ import com.ra.course.ams.airline.manag.system.entity.person.Pilot;
 import com.ra.course.ams.airline.manag.system.repository.Repository;
 import com.ra.course.ams.airline.manag.system.service.FlightInstanceService;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class FlightInstanceServiceImpl implements FlightInstanceService {
@@ -19,24 +20,36 @@ public class FlightInstanceServiceImpl implements FlightInstanceService {
 
     @Override
     public FlightInstance add(FlightInstance flightInstance) {
+        if (flightInstance == null) {
+            throw new NullPointerException("Cannot process add operation for null value argument.");
+        }
         return this.flightInstanceRepository.addInstance(flightInstance);
     }
 
     @Override
     public void updateStatus(FlightInstance flightInstance, FlightStatus status) {
+        if (flightInstance == null || status == null) {
+            throw new NullPointerException("Cannot process update status operation for null value argument.");
+        }
         flightInstance.setStatus(status);
         this.flightInstanceRepository.updateInstance(flightInstance);
     }
 
     @Override
     public boolean cancel(FlightInstance flightInstance) {
+        if (flightInstance == null) {
+            throw new NullPointerException("Cannot process cancel operation for null value argument.");
+        }
         this.flightInstanceRepository.removeInstance(flightInstance);
         return true;
     }
 
     @Override
     public boolean assignPilot(FlightInstance flightInstance, Pilot pilot) {
-        List<Pilot> pilots = flightInstance.getPilots();
+        if (flightInstance == null || pilot == null) {
+            throw new NullPointerException("Cannot process assign pilot operation for null value argument.");
+        }
+        List<Pilot> pilots = flightInstance.getPilots() != null ? flightInstance.getPilots() : new ArrayList<>();
         pilots.add(pilot);
         flightInstance.setPilots(pilots);
         this.flightInstanceRepository.updateInstance(flightInstance);
@@ -45,7 +58,10 @@ public class FlightInstanceServiceImpl implements FlightInstanceService {
 
     @Override
     public boolean assignCrew(FlightInstance flightInstance, Crew crew) {
-        List<Crew> crews = flightInstance.getCrews();
+        if (flightInstance == null || crew == null) {
+            throw new NullPointerException("Cannot process assign сrew operation for null value argument.");
+        }
+        List<Crew> crews = flightInstance.getCrews() != null ? flightInstance.getCrews() : new ArrayList<>();
         crews.add(crew);
         flightInstance.setCrews(crews);
         this.flightInstanceRepository.updateInstance(flightInstance);
